+++ lib/src/room_timeline.dart
@@ -142,6 +142,9 @@ class RoomTimeline extends Timeline {
       direction: Direction.b,
       historyCount: historyCount,
       filter: filter,
+      inMemoryOnly:
+          room.membership == Membership.leave &&
+          room.client.syncFilter.room?.includeLeave != true,
     );
     isRequestingHistory = false;
   }
@@ -262,6 +265,7 @@ class RoomTimeline extends Timeline {
     int historyCount = Room.defaultHistoryCount,
     required Direction direction,
     StateFilter? filter,
+    bool inMemoryOnly = false,
   }) async {
     onUpdate?.call();
 
@@ -310,7 +314,7 @@ class RoomTimeline extends Timeline {
         _fetchedAllDatabaseEvents = true;
         Logs().i('No more events found in the store. Request from server...');
 
-        if (isFragmentedTimeline) {
+        if (isFragmentedTimeline || inMemoryOnly) {
           await getRoomEvents(
             historyCount: historyCount,
             direction: direction,